requirements.txt
Pillow==2.9.0
praw==3.1.0
